Finalized shooting by distance and centering.
Now shooting by distance is working well, by the rate as fuction of
distance we have found. Centering is working, but not well. We
essentially overwrited the PID output with our own values, seperated,by
ranges. This is only a temporary solution, and we should complete it
*properly* for future use. Also integrated it into the rest of the robot
OI, so it can now be used natively, as a regular part of robot
operation.